+++ src/lib/types.ts
@@ -24,7 +24,9 @@ export namespace Constants {
     export const textual_preamble_size  = 1 + 10 + 1 + 8 + 1 + 4 + 1 + 1;
     export const textual_regexp  = /^\[[0-9]{4}-[0-9]{2}-[0-9]{2}T[0-9]{2}:[0-9]{2}:[0-9]{2}\+[0-9]{4}\]/;
 
-    export const issue_regexp = /^@?(scribejs|sjs),\s+(issue|pr)\s+(.*)$/;
+    export const issue_regexp    = /^@?(scribejs|sjs),\s+(issue|pr)\s+(.*)$/;
+
+    export const agenda_regexp   = /.* \-\- (.*) \-\-.*/;
 
     export const user_config_name = '.scribejs.json';
     export const user_ghid_file   = '.credentials.json';